Spawning items and ammo count
I now spawn ammo into the arena which can spawn in one of 12 locations. the player can also pick up this ammo to add ammo to the turret. The turret now only fires if the ammo count is greater than zero.

If an item trys to spawn where there is already an item it deletes one of the items meaning that the player cant just wait then collect all ammo at once.

Next is an ammo display on the screen and the ability to move the turret.