Ammo counter, turret movement and Hud Icons
I have made it so when the user is in the boss fight they switch between the fire boy and the water girl on the turret. water girl rotates the turret in the same way that water girl moves als switching to a camera on the crossbow in the scene

I have added an ammo count for the turret and fixed fixed an issue that I missed where the turret wasn't taking away ammo when the shot is fired.

Icons have been added to the hud for the user abilities, feather collection and arrows stored for the turret.

A health bar slide has been created but no code is currently behind it.