Player health
I have added code to the player health meaning now I can start to work on the ai damaging the player and the boss fight doing damage to the player if they are in the range of the fire.

The ability icons now switch if the player is on fire boy or water girl.