Boss fight damage with flames
the dragon now damages the player if they are in the zone that the dragon is in. some walls could be added to separate the arena for the user better.

Next step is to create a blast radius for the enemy ai. this will depend on how close the player is to the ai to how much damage the dammed ai will do.