ai explosion mechanic
I have added 3 triggers to the enemy ai so once the player is within the blast radius the ai will detonate reducing health off the player. I wanted to initialy do this using the range off the ai movement script however this did not work how I wanted it to. next stage is add something to show the user that the ai is going to detonate.