Fixed ai,placed all ai, placed all feathers,
I placed Ai around the arena so that they may challenge the user to navigate around the map

there are now 30 feathers placed around the area and they are well spread out.

Ai initialy only worked for 1 ai but now it works for all as the variables were only changing on the first of the ai and not all of them like it should of been so I placed the trigger bools on mechanics to fix this.